Switch all view-space coordinates on CanvasView from int to const double.

+++ api/objc/include/icCanvasManagerObjC/ICMCanvasView.h
@@ -10,15 +10,15 @@
 - (void)attachDrawing:(ICMDrawing*)drawing;
 
 - (void)drawWithContext:(cairo_t*)context intoRects:(cairo_rectangle_list_t*)rectList;
-- (void)setSizeWidth:(double)width andHeight:(double)height andUiScale:(double)ui_scale;
-- (void)setSizeUiScale:(double)ui_scale;
+- (void)setSizeWidth:(const double)width andHeight:(const double)height andUiScale:(const double)ui_scale;
+- (void)setSizeUiScale:(const double)ui_scale;
 - (void)getSizeWidth:(double*)out_width andHeight:(double*)out_height andUiScale:(double*)out_ui_scale;
-- (void)setScrollCenterX:(int)x andY:(int)y;
-- (void)setZoom:(int)vpixel_size;
+- (void)setScrollCenterX:(const double)x andY:(const double)y;
+- (void)setZoom:(const double)vpixel_size;
 
-- (void)mouseDownWithX:(int)x andY:(int)y andDeltaX:(int)deltaX andDeltaY:(int)deltaY;
-- (void)mouseDragWithX:(int)x andY:(int)y andDeltaX:(int)deltaX andDeltaY:(int)deltaY;
-- (void)mouseUpWithX:(int)x andY:(int)y andDeltaX:(int)deltaX andDeltaY:(int)deltaY;
+- (void)mouseDownWithX:(const double)x andY:(const double)y andDeltaX:(const double)deltaX andDeltaY:(const double)deltaY;
+- (void)mouseDragWithX:(const double)x andY:(const double)y andDeltaX:(const double)deltaX andDeltaY:(const double)deltaY;
+- (void)mouseUpWithX:(const double)x andY:(const double)y andDeltaX:(const double)deltaX andDeltaY:(const double)deltaY;
 
 - (void*)getWrappedObject;
